Refactor Donunt and Image app
