+++ game-basic/src/net/honux/game/basic/DonutBoard.java
@@ -1,11 +1,11 @@
-package net.honux.game;
+package net.honux.game.basic;
 
 import javax.swing.JPanel;
 import java.awt.*;
 import java.awt.geom.AffineTransform;
 import java.awt.geom.Ellipse2D;
 
-public class Board extends JPanel {
+public class DonutBoard extends JPanel {
 
     @Override
     public void paintComponent(Graphics g) {
@@ -31,13 +31,13 @@ public class Board extends JPanel {
         double w = size.getWidth();
         double h = size.getHeight();
 
-        Ellipse2D e = new Ellipse2D.Double(0, 0, 80, 130);
+        Ellipse2D e = new Ellipse2D.Double(0, 0, w / 6, h / 3);
         g2d.setStroke(new BasicStroke(1));
         g2d.setColor(Color.gray);
 
         for (double deg = 0; deg < 360; deg += 5) {
             AffineTransform at
-                    = AffineTransform.getTranslateInstance(w/2, h/2);
+                    = AffineTransform.getTranslateInstance(w / 2, h / 2);
             at.rotate(Math.toRadians(deg));
             g2d.draw(at.createTransformedShape(e));
         }
