+++ game-basic/src/net/honux/game/image/DrawImageExample.java
@@ -1,10 +1,7 @@
-package net.honux.game;
+package net.honux.game.image;
 
-import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
-import java.awt.image.BufferedImage;
-import java.io.File;
 import java.io.IOException;
 
 public class DrawImageExample extends JFrame {
@@ -33,5 +30,4 @@ public class DrawImageExample extends JFrame {
             app.setVisible(true);
         });
     }
-
 }
