+++ game-basic/src/net/honux/game/image/ImageBoard.java
@@ -1,4 +1,4 @@
-package net.honux.game;
+package net.honux.game.image;
 
 import javax.imageio.ImageIO;
 import javax.swing.*;
